Pass MYSQL root password to the backend app as an environment variable
This commit moves the root password specification to the docker-compose.yml file only.
When this is actually deployed in production, the passwords can be specified via
deployment specific environment variables from within Gitlab (so that they are not in
clear text). This way, development database always have the same password, but production
runs can (and should) have a different password.