Attempt infinite reconnections to the database backend if the connection is refused
On startup of the development server, it is necessary to bootstrap the MYSQL schema. This
ends up taking a few seconds and the datajoint backend will abort if it is unable to connect
to the database in this time. This commit adds an infinite while loop to continually poll
the MYSQL server until a connection is established.